refactor(policies): extracted into own files
