+++ pyproject.toml
@@ -122,3 +122,6 @@ version_scheme = "pep440"
 version_provider = "pep621"
 update_changelog_on_bump = true
 major_version_zero = true
+version_files = [
+  "src/vlagents/__init__.py:__version__",
+]
