+++ src/vlagents/policies/octo.py
@@ -1,0 +1,129 @@
+import base64
+import json
+import logging
+import os
+from collections import deque
+from dataclasses import dataclass, field
+from functools import partial, reduce
+from multiprocessing import resource_tracker, shared_memory
+from operator import getitem
+from pathlib import Path
+from typing import Any, Union
+
+import numpy as np
+import simplejpeg
+from PIL import Image
+from vlagents import register_agent
+from vlagents.policies.interface import Agent
+
+
+
+class OctoModel(Agent):
+    """
+    This model is trained with a window size of 2, predicting 7 dimensional actions 4 steps into the future.
+    Observations and tasks conform to the following spec:
+
+    Observations: {
+        image_primary: ('batch', 'history_window', 256, 256, 3),
+        image_wrist: ('batch', 'history_window', 128, 128, 3),
+    }
+    Tasks: {
+        image_primary: ('batch', 256, 256, 3),
+        image_wrist: ('batch', 128, 128, 3),
+        language_instruction: {
+            attention_mask: ('batch', 16),
+            input_ids: ('batch', 16),
+        },
+    }
+
+    At inference, you may pass in any subset of these observation and task keys, with a history window up to 2 timesteps.
+    """
+
+    def __init__(
+        self,
+        horizon: int = 2,
+        unnorm_key: list[str] | None = None,
+        default_checkpoint_path: str = "hf://rail-berkeley/octo-base-1.5",
+        **kwargs,
+    ) -> None:
+        # default window size is 2 in octo
+        # default unnorm is viola as it used the fr3
+        super().__init__(default_checkpoint_path=default_checkpoint_path, **kwargs)
+        self.horizon = horizon
+        if unnorm_key is None:
+            self.unnorm_key = []
+        else:
+            self.unnorm_key = unnorm_key
+
+        # log checkpoint path and step and kwargs
+        logging.info(f"checkpoint_path: {self.checkpoint_path}, checkpoint_step: {self.checkpoint_step}")
+        logging.info(f"horizon: {self.horizon}")
+        logging.info(f"unnorm_key: {self.unnorm_key}")
+        logging.info(f"kwargs: {kwargs}")
+        if self.checkpoint_path is None:
+            self.octo_path = self.default_checkpoint_path
+            logging.info(f"Using default checkpoint path: {self.octo_path}")
+            if self.checkpoint_step is not None:
+                logging.warning(
+                    "checkpoint_step should be None when using default path, ignoring checkpoint_step and setting it to None"
+                )
+                self.checkpoint_step = None
+            if self.unnorm_key != ["viola"]:
+                logging.warning(
+                    "unnorm_key should be ['viola'] when using default path, ignoring unnorm_key and setting it to ['viola']"
+                )
+                self.unnorm_key = ["viola"]
+        else:
+            self.octo_path = self.checkpoint_path
+            logging.info(f"Using custom checkpoint path: {self.octo_path} with checkpoint step: {self.checkpoint_step}")
+        logging.info(f"Using unnorm_key: {self.unnorm_key}")
+
+    def initialize(self):
+        os.environ["XLA_PYTHON_CLIENT_PREALLOCATE"] = (
+            "false"  # disable preallocation of memory in jax (might make it less efficient)
+        )
+        from octo.model.octo_model import OctoModel as _OctoModel
+        from octo.utils.train_callbacks import supply_rng
+
+        self.model = _OctoModel.load_pretrained(self.octo_path, self.checkpoint_step)
+
+        window_size = self.model.example_batch["observation"]["timestep_pad_mask"].shape[1]
+        if window_size < self.horizon:
+            logging.warning(
+                f"Horizon {self.horizon} is greater than the model's window size {window_size} with which the model has been trained with. "
+            )
+
+        self.trained_obs = self.model.example_batch["observation"].keys()
+
+        logging.info("==========================")
+        logging.info(self.model.dataset_statistics.keys())
+        self.policy_fn = supply_rng(
+            partial(
+                self.model.sample_actions,
+                unnormalization_statistics=reduce(getitem, self.unnorm_key, self.model.dataset_statistics)["action"],
+            ),
+        )
+
+    def act(self, obs: Obs) -> Act:
+        import jax
+        from octo.utils.gym_wrappers import stack_and_pad
+
+        super().act(obs)
+        robot_name, single_obs = self._require_single_arm(obs)
+        assert single_obs.cameras["rgb_side"].shape == (256, 256, 3), "wrong shape, use lanczos"
+        image_obs = {"image_primary": single_obs.cameras["rgb_side"]}
+        history = deque([image_obs] * self.horizon, maxlen=self.horizon)
+        full_obs = stack_and_pad(history, self.horizon)
+        task = self.model.create_tasks(texts=[obs.language_instruction or ""])
+
+        actions = self.policy_fn(
+            jax.tree_map(
+                lambda x: x[None],
+                full_obs,
+            ),
+            task,
+        )
+        action_chunk = np.asarray(actions[0, :, :], dtype=np.float32)
+        return self._chunk_act(robot_name, action_chunk[:, :-1], grippers=action_chunk[:, -1])
+
+register_agent("octo", OctoModel)
