+++ src/vlagents/policies/openvla.py
@@ -1,0 +1,100 @@
+import base64
+import json
+import logging
+import os
+from collections import deque
+from dataclasses import dataclass, field
+from functools import partial, reduce
+from multiprocessing import resource_tracker, shared_memory
+from operator import getitem
+from pathlib import Path
+from typing import Any, Union
+
+import numpy as np
+import simplejpeg
+from PIL import Image
+from vlagents import register_agent
+from vlagents.policies.interface import Agent
+
+class OpenVLAModel(Agent):
+    # === Utilities ===
+    SYSTEM_PROMPT = (
+        "A chat between a curious user and an artificial intelligence assistant. "
+        "The assistant gives helpful, detailed, and polite answers to the user's questions."
+    )
+
+    def __init__(
+        self,
+        attn_implementation: str,
+        device: str,
+        unnorm_key: str,
+        default_checkpoint_path: str = "openvla/openvla-7b",
+        **kwargs,
+    ) -> None:
+        super().__init__(default_checkpoint_path=default_checkpoint_path, **kwargs)
+        self.unnorm_key = unnorm_key
+        logging.info(f"Using unnorm_key: {self.unnorm_key}")
+        self.attn_implementation = attn_implementation
+        if self.checkpoint_step is None or self.checkpoint_path is None:
+            self.openvla_path = self.default_checkpoint_path
+            logging.info(f"Using default checkpoint path: {self.openvla_path}")
+            if self.unnorm_key != "viola":
+                logging.warning(
+                    "unnorm_key should be 'viola' when using default path, ignoring unnorm_key and setting it to 'viola'"
+                )
+                self.unnorm_key = "viola"
+        else:
+            self.openvla_path = self.checkpoint_path.format(checkpoint_step=self.checkpoint_step)
+            logging.info(
+                f"Using custom checkpoint path: {self.openvla_path} with checkpoint step: {self.checkpoint_step}"
+            )
+        self.device = device
+        self.attn_implementation = attn_implementation
+
+    def initialize(self):
+        import torch
+        from transformers import AutoModelForVision2Seq, AutoProcessor
+
+        self.device = torch.device(self.device) if torch.cuda.is_available() else torch.device("cpu")
+
+        # Load VLA Model using HF AutoClasses
+        self.processor = AutoProcessor.from_pretrained(self.openvla_path, trust_remote_code=True)
+        self.vla = AutoModelForVision2Seq.from_pretrained(
+            self.openvla_path,
+            attn_implementation=self.attn_implementation,
+            torch_dtype=torch.bfloat16,
+            low_cpu_mem_usage=True,
+            trust_remote_code=True,
+        ).to(self.device)
+        print("==========================")
+        print(self.vla.norm_stats.keys())
+
+        # [Hacky] Load Dataset Statistics from Disk (if passing a path to a fine-tuned model)
+        if os.path.isdir(self.openvla_path):
+            with open(Path(self.openvla_path) / "dataset_statistics.json", "r") as f:
+                self.vla.norm_stats = json.load(f)
+
+    def get_openvla_prompt(self, instruction: str, openvla_path: Union[str, Path]) -> str:
+        if "v01" in openvla_path:
+            return f"{self.SYSTEM_PROMPT} USER: What action should the robot take to {instruction.lower()}? ASSISTANT:"
+        else:
+            return f"In: What action should the robot take to {instruction.lower()}?\nOut:"
+
+    def act(self, obs: Obs) -> Act:
+        # no batch dimension here
+        import torch
+
+        super().act(obs)
+        robot_name, single_obs = self._require_single_arm(obs)
+        assert single_obs.cameras["rgb_side"].shape == (256, 256, 3), "wrong shape, use lanczos"
+        image = single_obs.cameras["rgb_side"]
+        unnorm_key = self.unnorm_key
+
+        prompt = self.get_openvla_prompt(obs.language_instruction or "", self.openvla_path)
+        inputs = self.processor(prompt, Image.fromarray(image).convert("RGB")).to(self.device, dtype=torch.bfloat16)
+        action = np.asarray(self.vla.predict_action(**inputs, unnorm_key=unnorm_key, do_sample=False), dtype=np.float32)
+        # OpenVLA returns a single step with gripper in the last dimension.
+        return self._single_step_act(robot_name, action[:-1], gripper=float(action[-1]))
+
+
+register_agent("openvla", OpenVLAModel)
